adding messagequeue_service
