+++ .gitignore
@@ -3,3 +3,4 @@
 /db/database.sqlite3
 /db/schema.sql
 /swagger.yaml
+.idea/
