+++ src/main.go
@@ -1,25 +1,16 @@
 package cicero
 
 import (
-	"context"
-	"database/sql"
-	"encoding/json"
-	"log"
-	"time"
-
 	nomad "github.com/hashicorp/nomad/api"
-	"github.com/liftbridge-io/go-liftbridge"
-	"github.com/pkg/errors"
+	"github.com/input-output-hk/cicero/src/model"
 	"github.com/uptrace/bun"
-	"github.com/uptrace/bun/dialect/sqlitedialect"
-	"github.com/uptrace/bun/driver/sqliteshim"
 )
 
 var DB *bun.DB
 var nomadClient *nomad.Client
 
 func Init() error {
-	openedDB, err := openDb()
+	openedDB, err := model.DBConnection()
 	if err != nil {
 		return err
 	}
@@ -33,63 +24,3 @@ func Init() error {
 	return nil
 }
 
-func openDb() (*bun.DB, error) {
-	sqldb, err := sql.Open(sqliteshim.ShimName, "db/database.sqlite3")
-	if err != nil {
-		return nil, errors.WithMessage(err, "While opening the DB")
-	}
-
-	db := bun.NewDB(sqldb, sqlitedialect.New())
-
-	return db, nil
-}
-
-func createStreams(logger *log.Logger, bridge liftbridge.Client, streamNames []string) error {
-	for _, streamName := range streamNames {
-		if err := bridge.CreateStream(
-			context.Background(),
-			streamName, streamName,
-			liftbridge.MaxReplication()); err != nil {
-			if err != liftbridge.ErrStreamExists {
-				if err != nil {
-					time.Sleep(1 * time.Second)
-					return errors.WithMessage(err, "Failed to Create NATS Stream")
-				}
-			}
-		} else {
-			logger.Printf("Created streams %s\n", streamName)
-		}
-	}
-
-	return nil
-}
-
-func publish(logger *log.Logger, bridge liftbridge.Client, stream, key string, msg map[string]interface{}) error {
-	err := createStreams(logger, bridge, []string{stream})
-	if err != nil {
-		return errors.WithMessage(err, "Before publishing message")
-	}
-
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-
-	enc, err := json.Marshal(msg)
-	if err != nil {
-		return errors.WithMessage(err, "Failed to encode JSON")
-	}
-
-	_, err = bridge.Publish(ctx, stream,
-		enc,
-		liftbridge.Key([]byte(key)),
-		liftbridge.PartitionByKey(),
-		liftbridge.AckPolicyAll(),
-	)
-
-	if err != nil {
-		return errors.WithMessage(err, "While publishing message")
-	}
-
-	logger.Printf("Published message to stream %s\n", stream)
-
-	return nil
-}
